+++ tests/conftest.py
@@ -1,8 +1,8 @@
 import pytest
 
 from roborock import HomeData, UserData
-from roborock.cloud_api import RoborockMqttClient
 from roborock.containers import DeviceData
+from roborock.version_1_apis.roborock_v1_mqtt_client import RoborockV1MqttClient
 from tests.mock_data import HOME_DATA_RAW, USER_DATA
 
 
@@ -14,6 +14,6 @@ def mqtt_client():
         device=home_data.devices[0],
         model=home_data.products[0].model,
     )
-    client = RoborockMqttClient(user_data, device_info)
+    client = RoborockV1MqttClient(user_data, device_info)
     yield client
     # Clean up any resources after the test
